GRIDEDIT-759 Fixed windows build failure

+++ libs/MeshKernel/include/MeshKernel/Polygon.hpp
@@ -101,7 +101,7 @@ namespace meshkernel
         /// The refined nodes are placed at equally spaced intervals for each polygon segment.
         /// The spacing of the refined nodes may differ between polygon segments.
         /// @return The points for the refined polygon
-        std::vector<Point> Refine(size_t startIndex, size_t endIndex, double refinementDistance) const;
+        std::vector<Point> Refine(UInt startIndex, UInt endIndex, double refinementDistance) const;
 
         /// @brief Refine the polygon
         /// @return The points for the refined polygon
